general md module started, and index cleaner
