+++ public/style.css
@@ -14,7 +14,7 @@ body{
 
 }
 .code{
-    color:#6a6a6a;
+    color:#6a9a6a;
     background-color:#4f4f4f; 
     width:100%; 
     font-size:65%;
